refactor(middleware): migrar middlewares para TypeScript
